refactor: 도메인 레이어 패키지 생성
