+++ src/main/java/com/capstone/moa/entity/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package com.capstone.moa;
+package com.capstone.moa.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
