+++ src/main/java/com/capstone/moa/entity/Job.java
@@ -1,4 +1,4 @@
-package com.capstone.moa;
+package com.capstone.moa.entity;
 
 public enum Job {
     STUDENT,
